fix unable to enter freq/chan

+++ ui/main.h
@@ -34,7 +34,7 @@ extern center_line_t g_center_line;
 #ifdef ENABLE_TX_AUDIO_BAR
 	bool UI_DisplayAudioBar(const bool now);
 #endif
-void UI_update_rssi(const int16_t rssi, const uint16_t glitch, const uint16_t noise, const int vfo);
+void UI_update_rssi(const int rssi, const unsigned int glitch, const unsigned int noise, const unsigned int vfo);
 void UI_DisplayMain(void);
 
 #endif
